requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pdfplumber==0.6.0
python-docx==0.8.11
pandas==2.1.1
spacy==3.6.1
nltk==3.9.0
pymongo==4.6.0
celery[redis]==5.4.0